requirements.txt
asgiref==3.4.1
boto3==1.20.2
botocore==1.23.3
celery==5.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-cors-headers==3.10.0
django-celery-results==2.1.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
s3transfer==0.5.0
pytest==6.2.5
python-decouple==3.6
psycopg2-binary==2.9.3